After last day of class

+++ pages/api/lists/[list]/index.js
@@ -1,5 +1,6 @@
 import { ObjectId } from "mongodb"
 import clientPromise from "../../../../lib/mongodb"
+import { unstable_getServerSession} from "next-auth"
 
 export default async function handler(req, res){
   switch(req.method){
@@ -22,6 +23,10 @@ export default async function handler(req, res){
     case 'PUT': // /api/lists/[list]
     case 'PATCH':
       try{
+        const session = await unstable_getServerSession(req, res, authOptions)
+        if(!session){
+          throw new Error("Unauthenticated request")
+        }
         const client = await clientPromise
     
         const db = client.db(process.env.DB_NAME)
@@ -29,25 +34,34 @@ export default async function handler(req, res){
         data.updatedAt = new Date()
         delete data['_id']
         let list_id = ObjectId(req.query.list)
-        await db.collection('lists').updateOne(
-          {_id: list_id},
-          {
-            $set: data
-          }
-        )
-        
-        res.status(200).json({success: true})
+        if (list.owner == session.user.email){
+          await db.collection('lists').updateOne(
+            {_id: list_id, owner: session.user.email },
+            {
+              $set: data
+            }
+          )
+          
+          res.status(200).json({success: true})
+        } else {
+          throw new Error("Unauthorized request")
+        }
       } catch (e) {
-        res.status(400).json({success: false})
+        res.status(400).json({success: false, message: e.message})
       }
       break
       case 'DELETE': // /api/lists/[list]
         try{
+          const session = await unstable_getServerSession(req, res, authOptions)
+          if(!session){
+            throw new Error("Unauthenticated request")
+          }
+
           const client = await clientPromise
       
           const db = client.db(process.env.DB_NAME)
           let list_id = ObjectId(req.query.list)
-          await db.collection('lists').deleteOne({_id: list_id} )
+          await db.collection('lists').deleteOne({_id: list_id, owner: session.user.email} )
           
           res.status(200).json({success: true})
         } catch (e) {
